Fix: wrap all Schema::table migrations with hasTable check

+++ database/migrations/2022_04_09_161150_add_wallet_point_columns_to_users_table.php
@@ -13,14 +13,16 @@ class AddWalletPointColumnsToUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            if (Schema::hasTable('users') && !Schema::hasColumn('users', 'wallet_balance')) {
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+            if (!Schema::hasColumn('users', 'wallet_balance')) {
                 $table->decimal('wallet_balance', 24, 3)->default(0);
             }
-            if (Schema::hasTable('users') && !Schema::hasColumn('users', 'loyalty_point')) {
+            if (!Schema::hasColumn('users', 'loyalty_point')) {
                 $table->decimal('loyalty_point', 24, 3)->default(0);
             }
-        });
+            });
+        }
     }
 
     /**
@@ -30,9 +32,11 @@ class AddWalletPointColumnsToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
             $table->dropColumn('wallet_balance');
             $table->dropColumn('loyalty_point');
-        });
+            });
+        }
     }
 }
