+++ database/migrations/2022_04_21_145207_add_column_to_modules_table.php
@@ -13,17 +13,13 @@ class AddColumnToModulesTable extends Migration
      */
     public function up()
     {
-        Schema::table('modules', function (Blueprint $table) {
-            if (Schema::hasTable('modules') && !Schema::hasColumn('modules', 'icon')) {
-                $table->string('icon',191)->nullable();
-            }
-            if (Schema::hasTable('modules') && !Schema::hasColumn('modules', 'theme_id')) {
-                $table->integer('theme_id')->default(1);
-            }
-            if (Schema::hasTable('modules') && !Schema::hasColumn('modules', 'description')) {
-                $table->text('description')->nullable();
-            }
-        });
+        if (Schema::hasTable('modules')) {
+            Schema::table('modules', function (Blueprint $table) {
+            $table->string('icon',191)->nullable();
+            $table->integer('theme_id')->default(1);
+            $table->text('description')->nullable();
+            });
+        }
     }
 
     /**
@@ -33,10 +29,12 @@ class AddColumnToModulesTable extends Migration
      */
     public function down()
     {
-        Schema::table('modules', function (Blueprint $table) {
+        if (Schema::hasTable('modules')) {
+            Schema::table('modules', function (Blueprint $table) {
             $table->dropColumn('icon');
             $table->dropColumn('theme_id');
             $table->dropColumn('description');
-        });
+            });
+        }
     }
 }
