+++ database/migrations/2022_05_12_170027_add_column_to_customer_addresses_table.php
@@ -13,17 +13,13 @@ class AddColumnToCustomerAddressesTable extends Migration
      */
     public function up()
     {
-        Schema::table('customer_addresses', function (Blueprint $table) {
-            if (Schema::hasTable('customer_addresses') && !Schema::hasColumn('customer_addresses', 'floor')) {
-                $table->string('floor')->nullable();
-            }
-            if (Schema::hasTable('customer_addresses') && !Schema::hasColumn('customer_addresses', 'road')) {
-                $table->string('road')->nullable();
-            }
-            if (Schema::hasTable('customer_addresses') && !Schema::hasColumn('customer_addresses', 'house')) {
-                $table->string('house')->nullable();
-            }
-        });
+        if (Schema::hasTable('customer_addresses')) {
+            Schema::table('customer_addresses', function (Blueprint $table) {
+            $table->string('floor')->nullable();
+            $table->string('road')->nullable();
+            $table->string('house')->nullable();
+            });
+        }
     }
 
     /**
@@ -33,10 +29,12 @@ class AddColumnToCustomerAddressesTable extends Migration
      */
     public function down()
     {
-        Schema::table('customer_addresses', function (Blueprint $table) {
+        if (Schema::hasTable('customer_addresses')) {
+            Schema::table('customer_addresses', function (Blueprint $table) {
             $table->dropColumn('floor');
             $table->dropColumn('road');
             $table->dropColumn('house');
-        });
+            });
+        }
     }
 }
