+++ database/migrations/2022_05_14_122603_add_dm_tips_column_to_order_transactions_table.php
@@ -13,11 +13,11 @@ class AddDmTipsColumnToOrderTransactionsTable extends Migration
      */
     public function up()
     {
-        Schema::table('order_transactions', function (Blueprint $table) {
-            if (Schema::hasTable('order_transactions') && !Schema::hasColumn('order_transactions', 'dm_tips')) {
-                $table->double('dm_tips', 24, 2)->default(0);
-            }
-        });
+        if (Schema::hasTable('order_transactions')) {
+            Schema::table('order_transactions', function (Blueprint $table) {
+            $table->double('dm_tips', 24, 2)->default(0);
+            });
+        }
     }
 
     /**
@@ -27,8 +27,10 @@ class AddDmTipsColumnToOrderTransactionsTable extends Migration
      */
     public function down()
     {
-        Schema::table('order_transactions', function (Blueprint $table) {
+        if (Schema::hasTable('order_transactions')) {
+            Schema::table('order_transactions', function (Blueprint $table) {
             $table->dropColumn('dm_tips');
-        });
+            });
+        }
     }
 }
