+++ database/migrations/2022_05_17_153333_add_ref_code_to_users_table.php
@@ -13,11 +13,11 @@ class AddRefCodeToUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            if (Schema::hasTable('users') && !Schema::hasColumn('users', 'ref_code')) {
-                $table->string('ref_code', 10)->nullable()->unique();
-            }
-        });
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+            $table->string('ref_code', 10)->nullable()->unique();
+            });
+        }
     }
 
     /**
@@ -27,8 +27,10 @@ class AddRefCodeToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
             $table->dropColumn('ref_code');
-        });
+            });
+        }
     }
 }
