+++ database/migrations/2022_10_25_153214_add_payment_method_columns_to_zones_table.php
@@ -13,14 +13,12 @@ class AddPaymentMethodColumnsToZonesTable extends Migration
      */
     public function up()
     {
-        Schema::table('zones', function (Blueprint $table) {
-            if (Schema::hasTable('zones') && !Schema::hasColumn('zones', 'cash_on_delivery')) {
-                $table->boolean('cash_on_delivery')->default(false);
-            }
-            if (Schema::hasTable('zones') && !Schema::hasColumn('zones', 'digital_payment')) {
-                $table->boolean('digital_payment')->default(false);
-            }
-        });
+        if (Schema::hasTable('zones')) {
+            Schema::table('zones', function (Blueprint $table) {
+            $table->boolean('cash_on_delivery')->default(false);
+            $table->boolean('digital_payment')->default(false);
+            });
+        }
     }
 
     /**
@@ -30,9 +28,11 @@ class AddPaymentMethodColumnsToZonesTable extends Migration
      */
     public function down()
     {
-        Schema::table('zones', function (Blueprint $table) {
+        if (Schema::hasTable('zones')) {
+            Schema::table('zones', function (Blueprint $table) {
             $table->dropColumn('cash_on_delivery');
             $table->dropColumn('digital_payment');
-        });
+            });
+        }
     }
 }
