+++ database/migrations/2022_12_20_104455_add_food_variations_column_to_items_table.php
@@ -13,11 +13,11 @@ class AddFoodVariationsColumnToItemsTable extends Migration
      */
     public function up()
     {
-        Schema::table('items', function (Blueprint $table) {
-            if (Schema::hasTable('items') && !Schema::hasColumn('items', 'food_variations')) {
-                $table->text('food_variations')->nullable();
-            }
-        });
+        if (Schema::hasTable('items')) {
+            Schema::table('items', function (Blueprint $table) {
+            $table->text('food_variations')->nullable();
+            });
+        }
     }
 
     /**
@@ -27,8 +27,10 @@ class AddFoodVariationsColumnToItemsTable extends Migration
      */
     public function down()
     {
-        Schema::table('items', function (Blueprint $table) {
+        if (Schema::hasTable('items')) {
+            Schema::table('items', function (Blueprint $table) {
             $table->dropColumn('food_variations');
-        });
+            });
+        }
     }
 }
