+++ database/migrations/2022_12_29_105321_add_maximum_cod_order_amount_column_to_module_zone_table.php
@@ -13,11 +13,11 @@ class AddMaximumCodOrderAmountColumnToModuleZoneTable extends Migration
      */
     public function up()
     {
-        Schema::table('module_zone', function (Blueprint $table) {
-            if (Schema::hasTable('module_zone') && !Schema::hasColumn('module_zone', 'maximum_cod_order_amount')) {
-                $table->double('maximum_cod_order_amount', 23, 2)->nullable();
-            }
-        });
+        if (Schema::hasTable('module_zone')) {
+            Schema::table('module_zone', function (Blueprint $table) {
+            $table->double('maximum_cod_order_amount', 23, 2)->nullable();
+            });
+        }
     }
 
     /**
@@ -27,8 +27,10 @@ class AddMaximumCodOrderAmountColumnToModuleZoneTable extends Migration
      */
     public function down()
     {
-        Schema::table('module_zone', function (Blueprint $table) {
+        if (Schema::hasTable('module_zone')) {
+            Schema::table('module_zone', function (Blueprint $table) {
             $table->dropColumn('maximum_cod_order_amount');
-        });
+            });
+        }
     }
 }
