+++ database/migrations/2023_01_03_093510_add_current_language_key_column_to_users_table.php
@@ -13,11 +13,11 @@ class AddCurrentLanguageKeyColumnToUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            if (Schema::hasTable('users') && !Schema::hasColumn('users', 'current_language_key')) {
-                $table->string('current_language_key')->default('en')->nullable();
-            }
-        });
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+            $table->string('current_language_key')->default('en')->nullable();
+            });
+        }
     }
 
     /**
@@ -27,8 +27,10 @@ class AddCurrentLanguageKeyColumnToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
             $table->dropColumn('current_language_key');
-        });
+            });
+        }
     }
 }
