+++ database/migrations/2023_01_23_144119_add_slug_to_item_campaigns_table.php
@@ -13,11 +13,11 @@ class AddSlugToItemCampaignsTable extends Migration
      */
     public function up()
     {
-        Schema::table('item_campaigns', function (Blueprint $table) {
-            if (Schema::hasTable('item_campaigns') && !Schema::hasColumn('item_campaigns', 'slug')) {
-                $table->string('slug')->nullable();
-            }
-        });
+        if (Schema::hasTable('item_campaigns')) {
+            Schema::table('item_campaigns', function (Blueprint $table) {
+            $table->string('slug')->nullable();
+            });
+        }
     }
 
     /**
@@ -27,8 +27,10 @@ class AddSlugToItemCampaignsTable extends Migration
      */
     public function down()
     {
-        Schema::table('item_campaigns', function (Blueprint $table) {
+        if (Schema::hasTable('item_campaigns')) {
+            Schema::table('item_campaigns', function (Blueprint $table) {
             $table->dropColumn('slug');
-        });
+            });
+        }
     }
 }
